build - adds a debian packaging maven profile
This allow to generate a debian package using the following command:

```
% mvn clean package deb:package -PdebianPackage -pl gateway
```

at the root of the repository.

Note: As opposed to what is done in georchestra/georchestra, we decided
to drop the `git clone datadir`, as people generally set up their infra
by doing a git clone by themselves of their datadir, so having a default
configuration provided in `/etc/georchestra` is not necessary.

Note2: the modification on the Java file is unrelated and is due to the
maven formatter plugin.