fix: removing TokenRelay filter by default
[TokenRelay](https://docs.spring.io/spring-cloud-gateway/reference/spring-cloud-gateway/gatewayfilter-factories/tokenrelay-factory.html)
is responsible for adding a Authorization header in the proxified HTTP
requests which can lead to weird behaviours with downstream webappS.

Considering that the geOrchestra applications will consume `sec-*` headers
for authnz, it does not really make sense to propagate the
`Authorization: bearer <oauth2 token>` to proxified webapps.

In fact, this behaviour lead to weird behaviours with some proxified
webapps, e.g. geocontrib replying to some HTTP requests with 403 because
it was not able to extract any info from the header.

It is still possible to use it by defining a custom filter on the route
which could be in the need for it, but it seems more reasonable to use
it as an "opt in" fashion instead of forcing it by default.